# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem ujian online (bank soal, jadwal, pengerjaan ujian per section, penilaian).",
        "title": "Ujian Backend API",
        "contact": {
            "name": "Tim Backend Ujian"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/kelas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "List kelas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan ID jurusan (uuid)",
                        "name": "id_jurusan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan tingkat",
                        "name": "tingkat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all kelas successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Buat kelas baru",
                "parameters": [
                    {
                        "description": "Data kelas",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.CreateKelasRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create kelas successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/kelas/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Detail kelas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kelas (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get kelas successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Kelas tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Update kelas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kelas (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data kelas",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.UpdateKelasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update kelas successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Hapus kelas (soft delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kelas (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete kelas successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal menghapus",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/kelas/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Restore kelas yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kelas (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore kelas successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal restore",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_kelas_dto.CreateKelasRequest": {
            "type": "object",
            "required": [
                "id_jurusan",
                "nama_kelas",
                "tingkat"
            ],
            "properties": {
                "id_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_kelas_dto.KelasListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_kelas_dto.KelasResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "string"
                },
                "nama_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_kelas_dto.UpdateKelasRequest": {
            "type": "object",
            "required": [
                "id_jurusan",
                "nama_kelas",
                "tingkat"
            ],
            "properties": {
                "id_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}